Documentation and script update
All GML Extended Scripts now use JSDoc-style comments compatible with docs_gm; consequently, there is now separate documenation for each script included with the source code! More documentation will be added to the Wiki, as well. Special thanks to Javier Mora for the help with this documentation.

To compliment this, some scripts have received slight updates. The most extensive are as follows:

* Constants representing the three possible return values of the Collision Checking scripts (and built-in functions) were added and incorporated into the scripts. These are INTERSECT_NONE (0), INTERSECT_INSIDE (1) and INTERSECT_OVERLAP (2).
* 'instance_on_screen' now accepts an 'ind' argument to reference instances that are not the caller.
* 'round_by' was changed to 'round_to' to match the other Real Number Functions