Update playable character animations and others
Used the Sonic 3 & Knuckles disassembly to update several animations so
they more closely resemble their origins. Also, updated MonitorIcon to
use timeline_set. Originally, this was left out of "GHZ boss update" as
a major refactoring of MonitorIcon was anticipated; it has since been
decided to leave MonitorIcon as is, besides this change.