Fix "time over" scenario (#18)
When the level time limit has expired, the current player character
should lose a life, and "Time Over" should appear on the screen (unless
it was the last life, in which case "Game Over" appears as normal.)
Until now, this was only partially implemented, and so reaching the
time limit did nothing. It is now fully functional.

This should fix "Global - Stage - Time Over does not occur #18".