Adjust conditions for landing on ceilings
The conditions for landing on ceiling terrain is strict. The terrain
must have the sticky flag set, it cannot be shallow or flat, and the
player cannot jump at it from the ground. These were set in place to
ensure the player could not break level design by jumping at low
ceilings and running along it. This isn't exactly accurate to how it
works in the classic games (and Sonic Mania,) but that was entirely on
purpose.
Nevertheless, the restriction of not being able to jump at the terrain
has been lifted. A typo was also fixed that prevented attaching to
ceilings from the left.